Improve main view behavior.
In preparation to actual saving of the expense.

+++ Expenses/Expenses.h
@@ -9,6 +9,6 @@
 @interface Category : NSObject
 
 @property (nonatomic) NSString *name;
-@property (nonatomic) uint32_t cid;
+@property (nonatomic) NSInteger cid;
 
 @end
